add corn job

+++ search_new.js
@@ -1,11 +1,8 @@
 var fs = require('fs');
 var readline = require('readline');
-const {
-  google
-} = require('googleapis');
-const {
-  OAuth2Client
-} = require('google-auth-library');
+const { google } = require('googleapis');
+const { OAuth2Client } = require('google-auth-library');
+var CronJob = require('cron').CronJob;
 
 // ----
 var MongoClient = require('mongodb').MongoClient;
@@ -49,13 +46,16 @@ fs.readFile('client_secret1.json', function processClientSecrets(err, content) {
     content
   };
   client_content = content;
-  // temp mark
-  searchVideos(obj);
-  // setInterval(() => {
-  //   searchVideos(obj);
-  // }, 20000);
-  // 測試 非同步連線 mongoDB
-  // mongoTestConnect();
+  new CronJob(
+    '0 0 12 * * *',
+    function() {
+      const d = new Date();
+      console.log('cornJobTime=>', d);
+      searchVideos(obj);
+    },
+    null,
+    true
+  );
 });
 
 // { findData, collectionName, callback }
@@ -64,7 +64,7 @@ async function mongoTestConnect() {
   var obj = {
     findData: {},
     collectionName: 'videos',
-    callback: function (doc) {
+    callback: function(doc) {
       console.log('find callback, doc is =>', doc);
     }
   };
@@ -86,7 +86,7 @@ function authorize(credentials, requestData, callback) {
   const oauth2Client = new OAuth2Client(clientId, clientSecret, redirectUrl);
   if (oauth2Data === '') {
     // Check if we have previously stored a token.
-    fs.readFile(TOKEN_PATH, function (err, token) {
+    fs.readFile(TOKEN_PATH, function(err, token) {
       if (err) {
         getNewToken(oauth2Client, requestData, callback);
       } else {
@@ -118,9 +118,9 @@ function getNewToken(oauth2Client, requestData, callback) {
     input: process.stdin,
     output: process.stdout
   });
-  rl.question('Enter the code from that page here: ', function (code) {
+  rl.question('Enter the code from that page here: ', function(code) {
     rl.close();
-    oauth2Client.getToken(code, function (err, token) {
+    oauth2Client.getToken(code, function(err, token) {
       if (err) {
         console.log('Error while trying to retrieve access token', err);
         return;
@@ -145,7 +145,7 @@ function storeToken(token) {
       throw err;
     }
   }
-  fs.writeFile(TOKEN_PATH, JSON.stringify(token), function (err, data) {
+  fs.writeFile(TOKEN_PATH, JSON.stringify(token), function(err, data) {
     if (err) throw err;
     console.log(data);
   });
@@ -212,7 +212,7 @@ function searchListByKeyword(auth, requestData) {
   var service = google.youtube('v3');
   var parameters = removeEmptyParameters(requestData['params']);
   parameters['auth'] = auth;
-  service.search.list(parameters, function (err, response) {
+  service.search.list(parameters, function(err, response) {
     if (err) {
       console.log('The API returned an error: ' + err);
       return;
@@ -226,11 +226,11 @@ let recordData = '';
 function searchVideos(option) {
   var keyword = option.keyword;
   // var callback = option.callback;
-  var callback = function (auth, requestData) {
+  var callback = function(auth, requestData) {
     var service = google.youtube('v3');
     var parameters = removeEmptyParameters(requestData['params']);
     parameters['auth'] = auth;
-    service.search.list(parameters, function (err, response) {
+    service.search.list(parameters, function(err, response) {
       if (err) {
         console.log('The API returned an error: ' + err);
         return;
@@ -252,9 +252,9 @@ function searchVideos(option) {
 
       // search videos function
       function recordCallback() {
-        nexPageToken = response.data.nextPageToken ?
-          response.data.nextPageToken :
-          '';
+        nexPageToken = response.data.nextPageToken
+          ? response.data.nextPageToken
+          : '';
         var channelData = '';
         var resData = response.data;
         var videos = [];
@@ -270,12 +270,14 @@ function searchVideos(option) {
               };
               channelHandler(channelData);
             }
-            if (newVideo !== '' &&
-              newVideo.videoId === item.id.videoId
-            ) {
+            if (newVideo !== '' && newVideo.videoId === item.id.videoId) {
               nexPageToken = '';
               newVideo = '';
-              console.log(`已經到了最新影片，停止新增！最新影片 videoId is ${item.id.videoId}`);
+              console.log(
+                `已經到了最新影片，停止新增！最新影片 videoId is ${
+                  item.id.videoId
+                }`
+              );
               break;
             } else videos.push(item);
           }
@@ -285,7 +287,7 @@ function searchVideos(option) {
           videos,
           channelId: channelData.channelId
         };
-        var _callback = function (obj) {
+        var _callback = function(obj) {
           insertVideosToUniqueTable(obj);
         };
         if (videos.length > 0) insertVideos(obj, _callback);
@@ -318,29 +320,32 @@ channelData = {
 */
 function insertChannel(channelData, callback) {
   // Connect to the db
-  callback = callback || function () {};
+  callback = callback || function() {};
   MongoClient.connect(
-    'mongodb://localhost:27017', {
+    'mongodb://localhost:27017',
+    {
       useNewUrlParser: true
     },
-    function (err, db) {
+    function(err, db) {
       if (err) throw err;
       var database = db.db('mydb');
       const collection = database.collection('channels');
-      var channels = [{
-        channelId: channelData.channelId,
-        channelTitle: channelData.channelTitle,
-        publishedAt: channelData.publishedAt
-      }];
+      var channels = [
+        {
+          channelId: channelData.channelId,
+          channelTitle: channelData.channelTitle,
+          publishedAt: channelData.publishedAt
+        }
+      ];
       // 寫入資料
-      collection.insertMany(channels, function (err, result) {
+      collection.insertMany(channels, function(err, result) {
         assert.equal(err, null);
         assert.equal(channels.length, result.result.n);
         assert.equal(channels.length, result.ops.length);
         console.log(
           'Inserted ' +
-          channels.length +
-          ' channels into the collection(channels)'
+            channels.length +
+            ' channels into the collection(channels)'
         );
         callback(result);
       });
@@ -352,12 +357,13 @@ function insertChannel(channelData, callback) {
 function channelHandler(channelData) {
   // Connect to the db
   MongoClient.connect(
-    'mongodb://localhost:27017', {
+    'mongodb://localhost:27017',
+    {
       useNewUrlParser: true
     },
-    function (err, db) {
+    function(err, db) {
       if (err) throw err;
-      var callbackFunction = function (res) {
+      var callbackFunction = function(res) {
         if (res.length === 0) {
           // 新增
           insertChannel(channelData);
@@ -377,11 +383,7 @@ function channelHandler(channelData) {
   );
 }
 // find channel
-const findChannel = function ({
-  db,
-  channelData,
-  callback
-}) {
+const findChannel = function({ db, channelData, callback }) {
   // Get the documents collection
   var database = db.db('mydb');
   const collection = database.collection('channels');
@@ -390,7 +392,7 @@ const findChannel = function ({
     .find({
       channelId: channelData.channelId
     })
-    .toArray(function (err, docs) {
+    .toArray(function(err, docs) {
       assert.equal(err, null);
       // console.log(docs);
       db.close(); //關閉連線
@@ -407,13 +409,14 @@ function insertVideos(option, callback) {
   var ch_id = option.channelId;
   var vds = option.videos;
   // Connect to the db
-  callback = callback || function () {};
+  callback = callback || function() {};
   if (vds.length > 0) {
     MongoClient.connect(
-      'mongodb://localhost:27017', {
+      'mongodb://localhost:27017',
+      {
         useNewUrlParser: true
       },
-      function (err, db) {
+      function(err, db) {
         if (err) throw err;
         var database = db.db('mydb');
         const collection = database.collection('videos');
@@ -432,7 +435,7 @@ function insertVideos(option, callback) {
         // console.log('videos=>', videos);
 
         // 寫入資料
-        collection.insertMany(videos, function (err, result) {
+        collection.insertMany(videos, function(err, result) {
           assert.equal(err, null);
           assert.equal(videos.length, result.result.n);
           assert.equal(videos.length, result.ops.length);
@@ -451,16 +454,17 @@ function insertVideos(option, callback) {
 function insertVideosToUniqueTable(data) {
   // Connect to the db
   MongoClient.connect(
-    'mongodb://localhost:27017', {
+    'mongodb://localhost:27017',
+    {
       useNewUrlParser: true
     },
-    function (err, db) {
+    function(err, db) {
       if (err) throw err;
       var database = db.db('mydb');
       var tableName = data[0].channelId + '_videos';
       const collection = database.collection(tableName);
       // 寫入資料
-      collection.insertMany(data, function (err, result) {
+      collection.insertMany(data, function(err, result) {
         assert.equal(err, null);
         assert.equal(data.length, result.result.n);
         assert.equal(data.length, result.ops.length);
@@ -492,11 +496,13 @@ function insertVideosToUniqueTable(data) {
 }
 // 檢查 紀錄檔資料，檢查最新的影片，
 function checkRecord(logData, _recordCallback) {
+  _recordCallback = _recordCallback || function() {};
   MongoClient.connect(
-    'mongodb://localhost:27017', {
+    'mongodb://localhost:27017',
+    {
       useNewUrlParser: true
     },
-    function (err, db) {
+    function(err, db) {
       if (err) throw err;
       var database = db.db('mydb');
       const collection = database.collection('record');
@@ -505,7 +511,7 @@ function checkRecord(logData, _recordCallback) {
         .find({
           videoId: logData.videoId
         })
-        .toArray(function (err, docs) {
+        .toArray(function(err, docs) {
           if (err) throw err;
           assert.equal(err, null);
           db.close(); //關閉連線
@@ -513,8 +519,13 @@ function checkRecord(logData, _recordCallback) {
             console.log(
               `Channel [${logData.channelTitle}] has new video, record log now!`
             );
-            recordLog(logData, _recordCallback);
+            recordLog({ logData, _recordCallback, isInsert: true });
           } else {
+            recordLog({
+              logData,
+              _recordCallback: function() {},
+              isInsert: false
+            });
             console.log(
               `Channel [${logData.channelTitle}] don't has new video!`
             );
@@ -525,25 +536,28 @@ function checkRecord(logData, _recordCallback) {
 }
 
 // 新增 紀錄檔資料
-function recordLog(logData, callback) {
-  callback = callback || function () {};
+function recordLog({ logData, callback, isInsert }) {
+  callback = callback || function() {};
   MongoClient.connect(
-    'mongodb://localhost:27017', {
+    'mongodb://localhost:27017',
+    {
       useNewUrlParser: true
     },
-    function (err, db) {
+    function(err, db) {
       if (err) throw err;
       var database = db.db('mydb');
       const collection = database.collection('record');
-      var record = [{
-        channelId: logData.channelId,
-        channelTitle: logData.channelTitle,
-        publishedAt: logData.publishedAt,
-        videoId: logData.videoId,
-        recordTime: new Date()
-      }];
+      var record = [
+        {
+          channelId: logData.channelId,
+          channelTitle: logData.channelTitle,
+          publishedAt: logData.publishedAt,
+          videoId: logData.videoId,
+          recordTime: new Date()
+        }
+      ];
       // 寫入資料
-      collection.insertMany(record, function (err, result) {
+      collection.insertMany(record, function(err, result) {
         assert.equal(err, null);
         assert.equal(record.length, result.result.n);
         assert.equal(record.length, result.ops.length);
@@ -551,44 +565,45 @@ function recordLog(logData, callback) {
           'Inserted ' + record.length + ' record into the collection(record)'
         );
       });
-      // 寫入自己的table
-      const uniq_collection = database.collection(
-        `${logData.channelId}_record`
-      );
-      // 去unique 那個資料夾把資料放入 newVideo ，再來清空資料表
-      uniq_collection
-        .find()
-        .toArray(function (err, docs) {
+      if (isInsert) {
+        // 寫入自己的table
+        const uniq_collection = database.collection(
+          `${logData.channelId}_record`
+        );
+        // 去unique 那個資料夾把資料放入 newVideo ，再來清空資料表
+        uniq_collection.find().toArray(function(err, docs) {
           if (err) throw err;
           assert.equal(err, null);
           if (docs.length != 0) newVideo = docs[0];
           console.log('newVideo => ', newVideo);
           uniq_collection
             .drop()
-            .then(function () {
+            .then(function() {
               // success
               assert.equal(err, null);
               // if (result) console.log('drop collection complete!');
             })
-            .catch(function () {
+            .catch(function() {
               // error handling
               // console.log('error handling');
             })
-            .finally(function () {
-              uniq_collection.insertMany(record, function (err, result) {
+            .finally(function() {
+              uniq_collection.insertMany(record, function(err, result) {
                 assert.equal(err, null);
                 assert.equal(record.length, result.result.n);
                 assert.equal(record.length, result.ops.length);
                 console.log(
                   `Inserted  ${record.length} record into the collection(${
-                          logData.channelId
-                        }_record})`
+                    logData.channelId
+                  }_record})`
                 );
               });
               callback(record);
               db.close(); //關閉連線
             });
-        })
+        });
+      }
+      db.close(); //關閉連線
     }
   );
 }
